[RCP-374] :recycle: 코드 구조 개선 및 마이페이지에서 내가 북마크, 좋아요한 페이지 이동 오류 수정
