+++ app/src/main/java/com/recipia/aos/ui/components/recipe/detail/comment/CommentsList.kt
@@ -1,0 +1,77 @@
+package com.recipia.aos.ui.components.recipe.detail.comment
+
+import android.annotation.SuppressLint
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.itemsIndexed
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+import com.recipia.aos.ui.components.HorizontalDivider
+import com.recipia.aos.ui.model.comment.CommentViewModel
+import kotlinx.coroutines.launch
+
+/**
+ * 댓글 리스트를 불러오는 컴포저
+ */
+@SuppressLint("CoroutineCreationDuringComposition")
+@Composable
+fun CommentsList(
+    commentViewModel: CommentViewModel,
+    recipeId: Long,
+    maxHeight: Dp
+) {
+    val commentsResponse by commentViewModel.comments.collectAsState()
+    val coroutineScope = rememberCoroutineScope()
+    val comments = commentsResponse?.content ?: emptyList()
+
+    Column {
+        // 고정된 헤더 부분
+        Text(
+            text = "댓글",
+            style = MaterialTheme.typography.titleMedium,
+            fontWeight = FontWeight.Bold,
+            modifier = Modifier
+                .fillMaxWidth() // 전체 너비를 채우도록 설정
+                .padding(start = 16.dp, bottom = 16.dp), // 양쪽에 패딩 적용
+        )
+
+        HorizontalDivider(
+            modifier = Modifier
+                .fillMaxWidth(), // 전체 너비를 채우도록 설정
+//                .padding(horizontal = 16.dp), // 양쪽에 패딩 적용
+            thickness = 0.5.dp, // 구분선의 두께 설정
+            color = Color(222, 226, 230) // 구분선의 색상 설정
+        )
+
+        Spacer(modifier = Modifier.height(12.dp))
+
+        // 스크롤 가능한 댓글 리스트 부분
+        LazyColumn(
+            modifier = Modifier.height(maxHeight)  // 여기에서 최대 높이를 적용합니다.
+        ) {
+            itemsIndexed(comments) { index, comment ->
+                CommentItem(comment)
+
+                // 마지막 항목이 렌더링되면 추가 데이터 로드
+                if (index == comments.size - 1) {
+                    coroutineScope.launch {
+                        commentViewModel.loadMoreComments(recipeId)
+                    }
+                }
+            }
+        }
+    }
+}
